Improved architecture and renaming of certain files to improve consistency
