+++ src/main/java/calculator/controller/components/RationalModeController.java
@@ -1,41 +1,39 @@
-package calculator.controller;
+package calculator.controller.components;
 
-import javafx.event.ActionEvent;
+import calculator.controller.MainViewController;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.scene.control.MenuItem;
-import javafx.scene.layout.GridPane;
+import javafx.scene.control.CheckMenuItem;
 import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class ComplexModeController implements Initializable {
+public class RationalModeController implements Initializable {
+    // TODO : Vérifier, utilisez public pour les méthodes de contrôleur qui sont appelées par le fichier FXML,
+    //  et utilisez private avec l'annotation @FXML pour les champs et méthodes qui sont utilisés uniquement
+    //  à l'intérieur de la classe de contrôleur.
 
-    private static final Logger logger = LoggerFactory.getLogger(ComplexModeController.class);
+    private static final Logger logger = LoggerFactory.getLogger(MainViewController.class);
 
     @FXML
-    private GridPane complexGrid;
+    private CheckMenuItem standardMode, rationalMode;
     @FXML
-    public Button btnExtension1, btnExtension2, btnExtension3, btnExtension4, btnExtension5; ;
+    private TextField display, expression;
     @FXML
-    private Button btnOpenParenthesis, btnCloseParenthesis, btnComma, btnClear, btnDivide, btnMultiply, btnMinus, btnPlus;
+    private Button btnOpenParenthesis, btnCloseParenthesis, btnComma, btnClear, btnFraction, btnMultiply, btnMinus, btnPlus;
     @FXML
     private Button btn0, btn1, btn2, btn3, btn4, btn5, btn6, btn7, btn8, btn9, btnEquals;
 
     private boolean resetDisplay = true;
-    private TextField display, expression;
 
     @Override
-    public void initialize(URL url, ResourceBundle resourceBundle) {}
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+        rationalMode.setSelected(true);
+    }
     /*public void appendToDisplay(String text) {
         if (resetDisplay) {
             expression.appendText(display.getText());
@@ -53,7 +51,7 @@ public class ComplexModeController implements Initializable {
         }
     }
 
-    public void onDivide() {
+    public void onFraction() {
         appendToDisplay("/");
     }
 
@@ -74,8 +72,6 @@ public class ComplexModeController implements Initializable {
         Button button = (Button) event.getSource();
         appendToDisplay(button.getText());
     }
-    
-    
 
     @FXML
     private void onStandardMode(ActionEvent actionEvent) {
@@ -137,16 +133,5 @@ public class ComplexModeController implements Initializable {
 
     public void displayDefaultZoom() {
         display.setStyle("-fx-font-size: 72px;");
-    }
-
-    public void setDisplayTextField(TextField display) {
-            this.display = display;
-    }
-
-    public void setExpressionTextField(TextField expression) {
-            this.expression = expression;
-    }
-
-    public void onExtension(ActionEvent actionEvent) {
     }*/
 }
